requirements.txt
streamlit>=1.20.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
python-dotenv>=1.0.0
langchain_groq
langchain-experimental
tabulate>=0.9.0